Math Practice
This program is geared towards elementary students that would seek to practice their math skills outside the class. The program generates two random numbers to generate an equation. When working with subtraction the program works to ensure there is no negatives and with division there is no dividing by zero.